refactor: replace with cli v2
