--- packages/lib/bin/cli-v2.js
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-"use strict";
-import("../dist/cli-v2.js");
