+++ packages/lib/bin/cli.js
