+++ packages/lib/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts", "src/cli.ts", "src/cli-v2.ts"],
+  entry: ["src/index.ts", "src/cli.ts"],
   format: ["esm"],
   dts: true,
 });
